+++ Source/Scene/ModelExperimental/ModelExperimental.js
@@ -1502,7 +1502,7 @@ ModelExperimental.prototype.destroyResources = function () {
  * The model can be a traditional glTF asset with a .gltf extension or a Binary glTF using the .glb extension.
  *
  * @param {Object} options Object with the following properties:
- * @param {String|Resource|Uint8Array|Object} options.gltf A Resource/URL to a glTF/glb file, a binary glTF buffer, or a JSON object containing the glTF contents
+ * @param {String|Resource} options.url The url to the .gltf or .glb file.
  * @param {String|Resource} [options.basePath=''] The base path that paths in the glTF JSON are relative to.
  * @param {Matrix4} [options.modelMatrix=Matrix4.IDENTITY] The 4x4 transformation matrix that transforms the model from model to world coordinates.
  * @param {Number} [options.scale=1.0] A uniform scale applied to this model.
@@ -1538,9 +1538,17 @@ ModelExperimental.prototype.destroyResources = function () {
  */
 ModelExperimental.fromGltf = function (options) {
   options = defaultValue(options, defaultValue.EMPTY_OBJECT);
+
   //>>includeStart('debug', pragmas.debug);
-  Check.defined("options.gltf", options.gltf);
+  if (!defined(options.url) && !defined(options.gltf)) {
+    throw new DeveloperError("options.url is required.");
+  }
   //>>includeEnd('debug');
+
+  // options.gltf is used internally for 3D Tiles. It can be a Resource, a URL
+  // to a glTF/glb file, a binary glTF buffer, or a JSON object containing the
+  // glTF contents.
+  const gltf = defaultValue(options.url, options.gltf);
 
   const loaderOptions = {
     releaseGltfJson: options.releaseGltfJson,
@@ -1549,8 +1557,6 @@ ModelExperimental.fromGltf = function (options) {
     forwardAxis: options.forwardAxis,
     loadIndicesForWireframe: options.enableDebugWireframe,
   };
-
-  const gltf = options.gltf;
 
   const basePath = defaultValue(options.basePath, "");
   const baseResource = Resource.createIfNeeded(basePath);
@@ -1564,7 +1570,7 @@ ModelExperimental.fromGltf = function (options) {
     loaderOptions.baseResource = baseResource;
     loaderOptions.gltfResource = baseResource;
   } else {
-    loaderOptions.gltfResource = Resource.createIfNeeded(options.gltf);
+    loaderOptions.gltfResource = Resource.createIfNeeded(gltf);
   }
 
   const loader = new GltfLoader(loaderOptions);